+++ src/app/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import {Route, Routes} from "react-router-dom";
-import {MainPage} from "pages/MainPage";
-import {Header} from "widgets/Header";
+import {Route, Routes} from 'react-router-dom'
+import {MainPage} from 'pages/MainPage'
+import {Header} from 'widgets/Header'
 
 
 function App() {
@@ -10,7 +10,7 @@ function App() {
             <Header/>
            
             <Routes>
-                <Route path={"/"} element={<MainPage/>}/>
+                <Route path={'/'} element={<MainPage/>}/>
             </Routes>
         </div>
     )
