+++ src/pages/MainPage/ui/MainPage.tsx
@@ -1,35 +1,36 @@
-import { memo } from "react";
-import styles from "./MainPage.module.scss";
-import classNames from "classnames";
-import { Container } from "shared/ui/Container/Container";
-import Text from "shared/ui/Text/Text";
-import { HandBook } from "widgets/HandBook/HandBook";
-import { Motorcycles } from "widgets/Motorcycles/Motorcycles";
-import { CommercialTransport } from "widgets/CommercialTransport/CommercialTransport";
+import {memo} from 'react'
+import styles from './MainPage.module.scss'
+import classNames from 'classnames'
+import {Container} from 'shared/ui/Container/Container'
+import Text from 'shared/ui/Text/Text'
+import {CommercialTransport} from 'widgets/CommercialTransport'
+import {Motorcycles} from 'widgets/Motorcycles'
+import {HandBook} from 'widgets/HandBook'
 
 interface MainPageProps {
-  className?: string;
+    className?: string;
 }
 
 const MainPage = (props: MainPageProps) => {
-  const { className } = props;
+    const {className} = props
 
     return (
-    <div className={classNames(styles.MainPage, {}, [className])}>
-      <Container>
-        <Text
-          title="Легковые автомобили в Москве"
-          description="7 532 предложения"
-        />
-      </Container>
-      <Container>
-        <CommercialTransport/>
-        <Motorcycles/>
-        <HandBook />
-      </Container>
-    </div>
-  );
+        <div className={classNames(styles.MainPage, {}, [className])}>
+            <Container>
+                <Text
+                    title="Легковые автомобили в Москве"
+                    description="7 532 предложения"
+                />
+            </Container>
 
-};
+            <Container>
+                <CommercialTransport/>
+                <Motorcycles/>
+                <HandBook/>
+            </Container>
+        </div>
+    )
 
-export default memo(MainPage);
+}
+
+export default memo(MainPage)
