+++ src/shared/ui/TileLink/TileLink.tsx
@@ -1,6 +1,7 @@
 import React, { memo, ReactNode, CSSProperties } from "react";
 import styles from "./TileLink.module.scss";
 import classNames from "classnames";
+import {AppLink} from "shared/ui/AppLink/AppLink"
 
 interface TileLinkProps {
   id?: string;
@@ -12,22 +13,22 @@ interface TileLinkProps {
 }
 
 function TileLink(props: TileLinkProps) {
-  const { className, children, urlLink, width, height } = props;
+    const { className, children, urlLink, width, height } = props;
 
-  const cls: CSSProperties = {
-    width,
-    height,
-  };
+    const cls: CSSProperties = {
+        width,
+        height,
+    };
 
-  return (
-    <a
-      className={classNames(className, styles.TileLink)}
-      style={cls}
-      href={urlLink}
-    >
-      <span className={classNames(styles.tileLinkTitle)}>{children}</span>
-    </a>
-  );
+    return (
+        <AppLink
+            className={classNames(className, styles.TileLink)}
+            style={cls}
+            to={urlLink}
+        >
+            <span className={classNames(styles.tileLinkTitle)}>{children}</span>
+        </AppLink>
+    );
 }
 
 export default memo(TileLink);
